refactor(front): move icon init logic to icons library
